fix(runtime): enforce the per-file ceiling where the bytes move
The ceiling lived only in staging, and the upload IPC takes the staged size as
optional — so the limit held only for callers that had actually staged. The
streamer now checks it itself, independent of whether a staged size was
supplied.

Both ceilings and their formatter move to runtime-import-limits so staging and
the streamer share one definition rather than drifting.
